pyproject.toml
[tool.poetry]
name = "little-observability"
version = "0.1.0"
description = ""
authors = ["Gary Bake <garybake@gmail.com>"]
readme = "README.md"
packages = [{include = "little_observability"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
